+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.0
 - Merged Iron's mana into Ars Nouveau mana
 - Merged Ars Nouveau spell power into Iron's spell power
+- Allowed Town Portal scrolls to be used in any dimension
 
 ## 1.1.1
 - Update 1.21.1 packs logo